+++ certstatus.go
@@ -1,29 +1,26 @@
 package main
 
 import (
-	"bytes"
-	"crypto"
 	"crypto/x509"
 	"encoding/pem"
 	"errors"
 	"flag"
 	"fmt"
-	"golang.org/x/crypto/ocsp"
 	"io"
 	"io/ioutil"
 	"net/http"
-	"net/url"
 	"os"
 )
 
 var (
-	errFailedToFetchOCSPResponse = errors.New("failed to fetch OCSP response")
-	errFailedToGetResource       = errors.New("failed to get resource")
-	errFailedToReadCertificate   = errors.New("failed to read certificate")
-	errFailedToReadResponseBody  = errors.New("failed to response body")
-	errNoCertificate             = errors.New("no certificate")
-	errNoIssuerCertificate       = errors.New("no issuer certificate")
-	errNoOCSPServersFound        = errors.New("no OCSP servers found")
+	errFailedToFetchOCSPResponse    = errors.New("failed to fetch OCSP response")
+	errFailedToGetResource          = errors.New("failed to get resource")
+	errFailedToReadCertificate      = errors.New("failed to read certificate")
+	errFailedToReadResponseBody     = errors.New("failed to response body")
+	errNoCertificate                = errors.New("no certificate")
+	errNoIssuerCertificate          = errors.New("no issuer certificate")
+	errNoOCSPServersFound           = errors.New("no OCSP servers found")
+	errNoCRLDistributionPointsFound = errors.New("no CRL distribution points found")
 
 	out    io.Writer  = os.Stdout // substituted during testing
 	client HttpClient = &http.Client{}
@@ -66,12 +63,20 @@ func printCertificateStatus(client HttpClient, path string) error {
 		return err
 	}
 
+	// OCSP
 	resp, err := getOCSPResponse(client, cert, issuer)
 	if err != nil {
 		return err
 	}
-
 	printStatusResponse(resp)
+
+	// CRL
+	crlresp, err := GetCRLResponse(client, cert)
+	//fmt.Print(crlresp.String())
+	if crlresp != nil {
+		fmt.Print(crlresp.String())
+	}
+
 	return nil
 }
 
@@ -110,56 +115,6 @@ func readCertificate(path string) (*x509.Certificate, error) {
 	return cert, nil
 }
 
-func getOCSPServer(cert *x509.Certificate) (string, error) {
-	ocspServers := cert.OCSPServer
-	if len(ocspServers) == 0 {
-		return "", errNoOCSPServersFound
-	}
-	return ocspServers[0], nil
-}
-
-func getOCSPResponse(client HttpClient, cert *x509.Certificate, issuer *x509.Certificate) (*ocsp.Response, error) {
-	ocspServer, err := getOCSPServer(cert)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "[error] %v\n", err)
-		os.Exit(1)
-	}
-
-	options := ocsp.RequestOptions{Hash: crypto.SHA1}
-	request, err := ocsp.CreateRequest(cert, issuer, &options)
-
-	url, err := url.Parse(ocspServer)
-
-	req, err := http.NewRequest("POST", ocspServer, bytes.NewBuffer(request))
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "[error] %v\n", err)
-		os.Exit(1)
-	}
-	req.Host = url.Hostname()
-	req.Header.Set("content-type", "application/ocsp-request")
-
-	response, err := client.Do(req)
-	if err != nil {
-		return nil, errFailedToFetchOCSPResponse
-	}
-	defer response.Body.Close()
-
-	body, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "[error] %v\n", err)
-		os.Exit(1)
-	}
-
-	parsedResponse, err := ocsp.ParseResponseForCert(body, cert, issuer)
-
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "[error] %v\n", err)
-		os.Exit(1)
-	}
-
-	return parsedResponse, nil
-}
-
 func getIssuerCertificate(client HttpClient, cert *x509.Certificate) (*x509.Certificate, error) {
 	var (
 		issCert *x509.Certificate
@@ -193,47 +148,4 @@ func getIssuerCertificate(client HttpClient, cert *x509.Certificate) (*x509.Cert
 	}
 
 	return issCert, nil
-}
-
-func printStatusResponse(resp *ocsp.Response) {
-	fmt.Fprintf(out, "Serial number: %s\n\n", resp.SerialNumber)
-	fmt.Fprintf(out, "Status: %s\n", statusMessage(resp.Status))
-
-	if resp.Status == ocsp.Revoked {
-		fmt.Fprintf(out, "Reason: %s\n", revocationReason(resp.RevocationReason))
-		fmt.Fprintf(out, "Revoked at: %s\n", resp.RevokedAt)
-	}
-
-	fmt.Fprintf(out, "\nProduced at: %s\n", resp.ProducedAt)
-	fmt.Fprintf(out, "This update: %s\n", resp.ThisUpdate)
-	fmt.Fprintf(out, "Next update: %s\n", resp.NextUpdate)
-}
-
-var (
-	statusMessages = map[int]string{
-		ocsp.Good:         "Good",
-		ocsp.Revoked:      "Revoked",
-		ocsp.ServerFailed: "Server failed",
-		ocsp.Unknown:      "Unknown",
-	}
-	revocationReasonMessages = map[int]string{
-		ocsp.Unspecified:          "Unspecified",
-		ocsp.KeyCompromise:        "Key compromise",
-		ocsp.CACompromise:         "CA compromise",
-		ocsp.AffiliationChanged:   "Affiliation changed",
-		ocsp.Superseded:           "Superseded",
-		ocsp.CessationOfOperation: "Cessation of operation",
-		ocsp.CertificateHold:      "Certificate hold",
-		ocsp.RemoveFromCRL:        "Remove from CRL",
-		ocsp.PrivilegeWithdrawn:   "Privilege withdrawn",
-		ocsp.AACompromise:         "AA compromise",
-	}
-)
-
-func statusMessage(code int) string {
-	return statusMessages[code]
-}
-
-func revocationReason(code int) string {
-	return revocationReasonMessages[code]
 }
